Done changes in sales and user flow

+++ userDetails.py
@@ -1,4 +1,6 @@
+import commonFunctions
 import pyodbc
+from prettytable import PrettyTable
 
 def getConnection():
     conn = pyodbc.connect('Driver={SQL Server};'
@@ -9,7 +11,8 @@ def getConnection():
     return cur, conn
 
 def displayMenu():
-    choice = int(input('Choose an option:\n 1. SqlScriptsAdd new Customer\n 2. Get Customer Details\n 3. Delete Customer Details'))
+    choice = int(input('1. Add new Customer\n2. Get Customer Details\n3. Delete Customer Details\n4. Quit \nChoose '
+                       'an option: '))
     cursor, conn = getConnection()
     if choice == 1:
         try:
@@ -21,13 +24,16 @@ def displayMenu():
                 max_id = 1
             
             name = input('Enter the name of Customer: ')
-            zip = int(input('Enter the ZIP de of customer: '))
+            zip = int(input('Enter the ZIP code of customer: '))
             tax_rate = int(input('Enter the tax percentage: '))
             email = input('Enter the email of Customer: ')
 
             query = f"insert into customerDetails values ({max_id},'{name}',{zip},{tax_rate},'{email}');"  
             cursor.execute(query)
             cursor.commit()
+            print("\n User registered successfully")
+            conn.close()
+            commonFunctions.displayMenu()
         except Exception as e:
             print(f'Failed to save customer details with exception: {e}')    
 
@@ -36,22 +42,36 @@ def displayMenu():
             customer_id = int(input('Enter customer ID: '))
             cursor.execute(f'select * from customerDetails where customerID = {customer_id};')
             results = cursor.fetchall()
-            print(results[0])
+
+            x = PrettyTable()
+
+            x.field_names = ["Customer ID", "Name", "ZIP", "Tax Rate", "Email"]
+            for result in results:
+                x.add_row(list(result))
+            print(x.get_string())
+            conn.close()
+            displayMenu()
         except Exception as e:
-            print(f'Failed to fetch customer details with exception: {e}')    
-    else:
+            print(f'Failed to fetch customer details with exception: {e}')
+    elif choice == 3:
         try:
             customer_id = int(input('Enter customer ID: '))
             cursor.execute(f'delete from customerDetails where customerID = {customer_id};')
+            print("Deleted user successfully")
             cursor.commit()
+            conn.close()
+            displayMenu()
         except Exception as e:
-            print(f'Failed to fetch customer details with exception: {e}')  
-    conn.close()
+            print(f'Failed to fetch customer details with exception: {e}')
+    else:
+        conn.close()
+        commonFunctions.displayMenu()
+
             
 
 
 def userDetails():
-    print("User Registration")
+    print("\nUser Registration\n")
     displayMenu()
